Fix code style and gyp files in cast to build cast_unittest
*.gypi should not be used for dependency directly. Renaming those files
to *.gyp. Also style fixes to make compilers happy.

This change is reland of r226264.

TBR=acolwell

Review URL: https://codereview.chromium.org/25544003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@226305 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 64582371e4e4988d9ea5292afa95c19072f3240b

+++ media/cast/video_receiver/codecs/vp8/vp8_decoder.cc
@@ -35,7 +35,7 @@ bool Vp8Decoder::Decode(const EncodedVideoFrame& input_image,
   vpx_image_t* img;
   if (vpx_codec_decode(decoder_.get(),
                        input_image.data.data(),
-                       input_image.data.size(),
+                       static_cast<unsigned int>(input_image.data.size()),
                        0,
                        1 /* real time*/)) {
     return false;
